backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "targetym-ai-backend"
version = "0.1.0"
description = "Backend API for Targetym AI - HR Analytics Platform"
authors = ["Targetym Team <team@targetym.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
redis = "^5.0.1"
celery = "^5.3.4"
httpx = "^0.25.2"
pandas = "^2.1.4"
numpy = "^1.25.2"
scikit-learn = "^1.3.2"
resend = "^0.7.0"
boto3 = "^1.34.0"
pillow = "^10.1.0"
jinja2 = "^3.1.2"
pydantic-settings = "^2.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
httpx = "^0.25.2"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
